# gaggiuino-local-profiler/go/internal/machines/handlers_control.go
package machines

import (
	"context"
	"encoding/json"
	"fmt"
	"log/slog"
	"net/http"
	"sync"

	"github.com/mxkissnr/gaggiuino-local-profiler/go/internal/httputil"
	"github.com/mxkissnr/gaggiuino-local-profiler/go/internal/machines/proto"
)

// This file is the #597 Gaggiuino settings/control proxy (settings
// read/write, opmode/tare/service-test, active-profile persistence,
// firmware OTA, live sensor/system state). Every route here is gated by
// requireSettingsProxySupport.

var gaggiuinoSettingsCategories = map[string]bool{
	"boiler": true, "system": true, "display": true, "scales": true, "led": true, "theme": true,
}

func (h *Handlers) registerControlRoutes(mux *http.ServeMux) {
	mux.HandleFunc("GET /api/machine/settings", h.getSettings)
	// Registered before the /{category} route below so this exact path
	// always wins — Go's ServeMux already prefers the more specific
	// literal pattern regardless of registration order (see
	// shots/handlers.go's header comment on the same non-issue), but
	// registered in this order for readability.
	mux.HandleFunc("POST /api/machine/settings/save", h.saveSettings)
	mux.HandleFunc("POST /api/machine/settings/{category}", h.updateSettings)

	mux.HandleFunc("POST /api/machine/opmode", h.setOperationMode)
	mux.HandleFunc("POST /api/machine/tare", h.tare)
	mux.HandleFunc("POST /api/machine/service-test", h.serviceTest)

	mux.HandleFunc("POST /api/machine/profile/save", h.saveActiveProfile)

	mux.HandleFunc("GET /api/machine/firmware/progress", h.firmwareProgress)
	mux.HandleFunc("POST /api/machine/firmware/update", h.triggerFirmwareUpdate)
	mux.HandleFunc("GET /api/machine/firmware/version", h.firmwareVersion)

	mux.HandleFunc("GET /api/machine/live", h.machineLive)
}

// resolveWithAdapter performs the repeated resolve-machine + get-adapter
// pair every route in this file (and handlers_profiles.go) opens with.
func (h *Handlers) resolveWithAdapter(w http.ResponseWriter, machineID *int64) (*Machine, Adapter, bool) {
	machine, err := h.registry.ResolveMachine(machineID)
	if err != nil {
		internalError(w, err)
		return nil, nil, false
	}
	adapter, err := h.GetAdapter(machine)
	if err != nil {
		internalError(w, err)
		return nil, nil, false
	}
	return machine, adapter, true
}

func (h *Handlers) getSettings(w http.ResponseWriter, r *http.Request) {
	machine, adapter, ok := h.resolveWithAdapter(w, queryMachineID(r))
	if !ok {
		return
	}
	if !requireSettingsProxySupport(w, adapter, machine) {
		return
	}
	category := r.URL.Query().Get("category")
	if category != "" && !gaggiuinoSettingsCategories[category] && category != "versions" {
		writeError(w, http.StatusBadRequest, "unknown settings category: "+category)
		return
	}
	settings, err := adapter.GetSettings(r.Context(), machine, category)
	if err != nil {
		writeError(w, http.StatusBadGateway, err.Error())
		return
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write(settings)
}

func (h *Handlers) saveSettings(w http.ResponseWriter, r *http.Request) {
	var body struct {
		MachineID *int64 `json:"machineId"`
	}
	if !decodeJSONBody(w, r, &body) {
		return
	}
	machine, adapter, ok := h.resolveWithAdapter(w, body.MachineID)
	if !ok {
		return
	}
	if !requireSettingsProxySupport(w, adapter, machine) {
		return
	}
	if err := adapter.SaveSettings(r.Context(), machine); err != nil {
		writeError(w, http.StatusBadGateway, err.Error())
		return
	}
	writeJSON(w, http.StatusOK, map[string]bool{"ok": true})
}

func (h *Handlers) updateSettings(w http.ResponseWriter, r *http.Request) {
	rawBody, ok := readRawJSONBody(w, r)
	if !ok {
		return
	}
	var body struct {
		MachineID *int64 `json:"machineId"`
	}
	_ = json.Unmarshal(rawBody, &body) // best-effort: only used to resolve the machine

	machine, adapter, ok := h.resolveWithAdapter(w, body.MachineID)
	if !ok {
		return
	}
	if !requireSettingsProxySupport(w, adapter, machine) {
		return
	}
	category := r.PathValue("category")
	if !gaggiuinoSettingsCategories[category] {
		writeError(w, http.StatusBadRequest, "unknown or read-only settings category: "+category)
		return
	}
	if err := ValidateSettingsPayload(rawBody); err != nil {
		writeError(w, http.StatusBadRequest, err.Error())
		return
	}
	result, err := adapter.UpdateSettings(r.Context(), machine, category, rawBody)
	if err != nil {
		writeError(w, http.StatusBadGateway, err.Error())
		return
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write(result)
}

func (h *Handlers) setOperationMode(w http.ResponseWriter, r *http.Request) {
	var body struct {
		Mode      proto.OperationMode `json:"mode"`
		MachineID *int64              `json:"machineId"`
	}
	if !decodeJSONBody(w, r, &body) {
		return
	}
	machine, adapter, ok := h.resolveWithAdapter(w, body.MachineID)
	if !ok {
		return
	}
	if !requireSettingsProxySupport(w, adapter, machine) {
		return
	}
	if err := ValidateOperationMode(body.Mode); err != nil {
		writeError(w, http.StatusBadRequest, "invalid mode: "+err.Error())
		return
	}
	if err := adapter.SetOperationMode(r.Context(), machine, body.Mode); err != nil {
		writeError(w, http.StatusBadGateway, err.Error())
		return
	}
	writeJSON(w, http.StatusOK, map[string]bool{"ok": true})
}

func (h *Handlers) tare(w http.ResponseWriter, r *http.Request) {
	var body struct {
		MachineID *int64 `json:"machineId"`
	}
	if !decodeJSONBody(w, r, &body) {
		return
	}
	machine, adapter, ok := h.resolveWithAdapter(w, body.MachineID)
	if !ok {
		return
	}
	if !requireSettingsProxySupport(w, adapter, machine) {
		return
	}
	if err := adapter.Tare(r.Context(), machine); err != nil {
		writeError(w, http.StatusBadGateway, err.Error())
		return
	}
	writeJSON(w, http.StatusOK, map[string]bool{"ok": true})
}

func (h *Handlers) serviceTest(w http.ResponseWriter, r *http.Request) {
	var body struct {
		Peripheral proto.ServiceTestPeripheral `json:"peripheral"`
		MachineID  *int64                      `json:"machineId"`
	}
	if !decodeJSONBody(w, r, &body) {
		return
	}
	machine, adapter, ok := h.resolveWithAdapter(w, body.MachineID)
	if !ok {
		return
	}
	if !requireSettingsProxySupport(w, adapter, machine) {
		return
	}
	if err := ValidateServiceTestPeripheral(body.Peripheral); err != nil {
		writeError(w, http.StatusBadRequest, "invalid peripheral: "+err.Error())
		return
	}
	if err := adapter.ServiceTest(r.Context(), machine, body.Peripheral); err != nil {
		writeError(w, http.StatusBadGateway, err.Error())
		return
	}
	writeJSON(w, http.StatusOK, map[string]bool{"ok": true})
}

func (h *Handlers) saveActiveProfile(w http.ResponseWriter, r *http.Request) {
	var body struct {
		MachineID *int64 `json:"machineId"`
	}
	if !decodeJSONBody(w, r, &body) {
		return
	}
	machine, adapter, ok := h.resolveWithAdapter(w, body.MachineID)
	if !ok {
		return
	}
	if !requireSettingsProxySupport(w, adapter, machine) {
		return
	}
	if err := adapter.SaveActiveProfile(r.Context(), machine); err != nil {
		writeError(w, http.StatusBadGateway, err.Error())
		return
	}
	writeJSON(w, http.StatusOK, map[string]bool{"ok": true})
}

func (h *Handlers) firmwareProgress(w http.ResponseWriter, r *http.Request) {
	machine, adapter, ok := h.resolveWithAdapter(w, queryMachineID(r))
	if !ok {
		return
	}
	if !requireSettingsProxySupport(w, adapter, machine) {
		return
	}
	progress, err := adapter.GetFirmwareProgress(r.Context(), machine)
	if err != nil {
		writeError(w, http.StatusBadGateway, err.Error())
		return
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write(progress)
}

func (h *Handlers) triggerFirmwareUpdate(w http.ResponseWriter, r *http.Request) {
	var body struct {
		MachineID *int64 `json:"machineId"`
	}
	if !decodeJSONBody(w, r, &body) {
		return
	}
	machine, adapter, ok := h.resolveWithAdapter(w, body.MachineID)
	if !ok {
		return
	}
	if !requireSettingsProxySupport(w, adapter, machine) {
		return
	}
	// #1136 follow-up: resolve the installed/target version BEFORE triggering
	// the update -- a successful trigger reboots the machine, so after that
	// call neither `versions` nor `system` can be read reliably. Only done
	// when a maintenance-log hook is wired, and entirely best-effort: any
	// error or panic inside the lookup degrades to an empty string for that
	// side and never touches the update or this response.
	var from, to string
	if h.onFirmwareUpdate != nil {
		httputil.SafeCall("machines: firmware update version lookup", func() {
			from, to = h.firmwareFromTo(r.Context(), machine, adapter)
		})
	}
	result, err := adapter.TriggerFirmwareUpdate(r.Context(), machine)
	if err != nil {
		writeError(w, http.StatusBadGateway, err.Error())
		return
	}
	// #1136: record the update in the machine's maintenance log. Only on
	// success (the failure path above already returned), and never let the
	// hook change this response -- its error is logged, and a panic inside
	// it is contained by SafeCall.
	if h.onFirmwareUpdate != nil {
		var cbErr error
		httputil.SafeCall("machines: firmware update maintenance log", func() {
			cbErr = h.onFirmwareUpdate(machine, from, to)
		})
		if cbErr != nil {
			slog.Warn("firmware update: maintenance log callback failed", "err", cbErr)
		}
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write(result)
}

// firmwareFromTo best-effort resolves the version being replaced (from: the
// machine's installed `versions`.coreVersion) and the version it is updating
// to (to: the latest release on the machine's `system`.releaseChannel),
// mirroring the reads firmwareVersion above already does. Both reads are
// wrapped in SafeCall and swallow every error/panic, so a missing or
// unreachable machine simply yields an empty side rather than affecting the
// update; to stays empty when the release lookup fails or reports nothing.
func (h *Handlers) firmwareFromTo(ctx context.Context, machine *Machine, adapter Adapter) (from, to string) {
	var versionsRaw json.RawMessage
	httputil.SafeCall("machines: firmware update from-version fetch", func() {
		versionsRaw, _ = adapter.GetSettings(ctx, machine, "versions")
	})
	var versions struct {
		CoreVersion *string `json:"coreVersion"`
	}
	if len(versionsRaw) > 0 {
		_ = json.Unmarshal(versionsRaw, &versions)
	}
	if versions.CoreVersion != nil {
		from = *versions.CoreVersion
	}

	var systemRaw json.RawMessage
	httputil.SafeCall("machines: firmware update system-settings fetch", func() {
		systemRaw, _ = adapter.GetSettings(ctx, machine, "system")
	})
	var system map[string]any
	if len(systemRaw) > 0 {
		_ = json.Unmarshal(systemRaw, &system)
	}
	channel := ParseReleaseChannel(system["releaseChannel"])
	httputil.SafeCall("machines: firmware update latest-release lookup", func() {
		latest, err := h.firmware.GetLatestFirmwareRelease(ctx, channel)
		if err == nil && latest != nil {
			to = latest.Hash
		}
	})
	return from, to
}

// firmwareVersion serves GET /api/machine/firmware/version (#620).
func (h *Handlers) firmwareVersion(w http.ResponseWriter, r *http.Request) {
	machine, adapter, ok := h.resolveWithAdapter(w, queryMachineID(r))
	if !ok {
		return
	}
	if !requireSettingsProxySupport(w, adapter, machine) {
		return
	}
	// The "versions" and "system" settings reads are independent (#901 code
	// review), so fetch them concurrently instead of paying two round-trips
	// back to back.
	var versionsRaw, systemRaw json.RawMessage
	var versionsErr, systemErr error
	var versionsPanicked bool
	var wg sync.WaitGroup
	wg.Add(2)
	go func() {
		defer wg.Done()
		if httputil.SafeCall("machines: firmware version fetch", func() {
			versionsRaw, versionsErr = adapter.GetSettings(r.Context(), machine, "versions")
		}) {
			versionsErr = fmt.Errorf("internal error fetching versions settings")
			versionsPanicked = true
		}
	}()
	go func() {
		defer wg.Done()
		if httputil.SafeCall("machines: firmware version fetch", func() {
			systemRaw, systemErr = adapter.GetSettings(r.Context(), machine, "system")
		}) {
			systemErr = fmt.Errorf("internal error fetching system settings")
		}
	}()
	wg.Wait()
	// #1037: a recovered panic is a genuine backend bug, not a routine
	// offline machine -- keep surfacing that as 502 (see
	// TestFirmwareVersion_PanicDuringSettingsFetchReturns502).
	//
	// #1046: every other `versions` fetch failure (the ordinary case: the
	// machine's host is simply unreachable) used to also 502 here, back when
	// this endpoint only ever fired from an explicit "open Edit form" click.
	// #1046 moved the call onto every Gaggiuino row's automatic list-render
	// load, so an offline machine now hits this on every Settings open --
	// same tolerant-of-unreachable shape testMachine() (handlers_registry.go)
	// already uses for "Verbindung testen", degrading to a 200 with nulled-
	// out fields instead of a 502 the browser itself logs as a console error
	// regardless of any client-side .catch().
	if versionsErr != nil {
		if versionsPanicked {
			writeError(w, http.StatusBadGateway, versionsErr.Error())
			return
		}
		slog.Warn("firmware version: versions settings fetch failed, reporting unknown", "err", versionsErr)
		writeJSON(w, http.StatusOK, map[string]any{
			"installed": nil, "latest": nil, "updateAvailable": false, "releaseUrl": nil,
		})
		return
	}
	var versions struct {
		CoreVersion *string `json:"coreVersion"`
	}
	_ = json.Unmarshal(versionsRaw, &versions)
	var system map[string]any
	if systemErr != nil {
		slog.Warn("firmware version: system settings fetch failed, using default release channel", "err", systemErr)
	} else {
		_ = json.Unmarshal(systemRaw, &system)
	}

	channel := ParseReleaseChannel(system["releaseChannel"])
	latest, err := h.firmware.GetLatestFirmwareRelease(r.Context(), channel)
	if err != nil {
		slog.Warn("firmware version: latest-release lookup failed, reporting installed version only", "err", err)
		latest = nil
	}

	installed := versions.CoreVersion
	updateAvailable := installed != nil && *installed != "" && latest != nil && *installed != latest.Hash
	resp := map[string]any{
		"installed":       nullOr(installed),
		"latest":          nil,
		"updateAvailable": updateAvailable,
		"releaseUrl":      nil,
	}
	if latest != nil {
		resp["latest"] = latest.Hash
		resp["releaseUrl"] = latest.ReleaseURL
	}
	writeJSON(w, http.StatusOK, resp)
}

func nullOr(s *string) any {
	if s == nil {
		return nil
	}
	return *s
}

// machineLive serves GET /api/machine/live: latest cached live sensor/
// system-state pushes from the machine's persistent WebSocket session
// (live.go) — null until the first push arrives.
func (h *Handlers) machineLive(w http.ResponseWriter, r *http.Request) {
	machine, adapter, ok := h.resolveWithAdapter(w, queryMachineID(r))
	if !ok {
		return
	}
	if !requireSettingsProxySupport(w, adapter, machine) {
		return
	}
	sensorSnap, err := adapter.GetLiveSensorSnapshot(r.Context(), machine)
	if err != nil {
		writeError(w, http.StatusBadGateway, err.Error())
		return
	}
	sysState, err := adapter.GetLiveSystemState(r.Context(), machine)
	if err != nil {
		writeError(w, http.StatusBadGateway, err.Error())
		return
	}
	writeJSON(w, http.StatusOK, map[string]any{
		"sensorSnap": sensorSnap,
		"sysState":   sysState,
	})
}