# gaggiuino-local-profiler/go/internal/system/openapi.yaml
openapi: 3.0.3
info:
  title: GLP REST API
  version: 2.0.0
  description: >
    REST API for the **Gaggiuino Local Profiler** Home Assistant add-on.
    Provides shot storage, coffee library, maintenance tracking, a barista
    orders system, and live machine data.
  contact:
    url: https://github.com/mxkissnr/gaggiuino-local-profiler
servers:
  - url: http://localhost:8099
    description: Direct port (integration, cards)
  - url: /api/hassio_ingress/gaggiuino_local_profiler
    description: HA Ingress (browser — auth via HA session, no token needed)
security:
  - tokenAuth: []
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: X-GLP-Token
      description: >
        64-char hex token, auto-generated on first start, stored in
        `/data/api_token.txt`. Retrieve from `GET /api/status`.
        Requests via HA Ingress bypass this check (HA session is sufficient).
        `GET /api/status` is always public.
    tokenQueryAuth:
      type: apiKey
      in: query
      name: token
      description: >
        Same token as `tokenAuth`, accepted as a `?token=` query parameter
        instead of the `X-GLP-Token` header. Only honored on `GET
        /api/events`, since a browser-native EventSource connection can't
        set custom headers — every other route requires the header.
  schemas:
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, example: true }
...
    MachineSystemSettings:
      type: object
      description: >
        GET /api/machine/settings — the machine's own settings object, passed
        through from its GET /api/settings/{category}. Only releaseChannel is
        read by the app today, so the rest stays an open object.
      properties:
        releaseChannel: { type: integer }
      additionalProperties: true
...
    FirmwareProgress:
      type: object
      description: >
        GET /api/machine/firmware/progress — the machine's own OTA progress
        payload, passed through verbatim (progress/status/type today).
      properties:
        status: { type: string }
      additionalProperties: true
...
paths:
...
  /api/machine/settings:
    get:
      tags: [MachineControl]
      summary: Read the machine's settings (all categories, or one via ?category=)
      parameters:
        - { name: machineId, in: query, schema: { type: integer } }
        - { name: category, in: query, schema: { type: string, enum: [boiler, system, display, scales, led, theme, versions] } }
      responses:
        "200":
          description: Settings object (shape mirrors the machine's own GET /api/settings/* response)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MachineSystemSettings' }
        "400": { description: unknown category }
        "501": { description: machine type does not support the settings proxy }
        "502": { description: Machine unreachable or returned an error }
  /api/machine/settings/save:
    post:
      tags: [MachineControl]
      summary: Persist whatever settings are currently applied in RAM to flash (c_save_settings over WebSocket)
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { machineId: { type: integer } } }
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
        "501": { description: machine type does not support the settings proxy }
        "502": { description: Machine unreachable, timed out, or rejected the command }
  /api/machine/settings/{category}:
    post:
      tags: [MachineControl]
      summary: Update one settings category — proxies the machine's own POST /api/settings/{category}, which applies and persists in one call
      parameters:
        - { name: category, in: path, required: true, schema: { type: string, enum: [boiler, system, display, scales, led, theme] } }
      requestBody:
        content:
          application/json:
            schema: { type: object, description: "Full category payload (all fields from the matching GET response) — passed through as-is; machineId may be included alongside it" }
      responses:
        "200": { description: "{ success: true }" }
        "400": { description: unknown/read-only category, or invalid payload }
        "501": { description: machine type does not support the settings proxy }
        "502": { description: Machine unreachable or returned an error }
  /api/machine/opmode:
    post:
      tags: [MachineControl]
      summary: Switch operation mode (c_opmode over WebSocket)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                mode: { description: "OperationModeDto enum name or numeric value — BREW_MANUAL is rejected (live-verified no-op while idle)", oneOf: [{ type: string, enum: [BREW_AUTO, FLUSH, DESCALE, STEAM, FLUSH_AUTO, HOT_WATER, HOME] }, { type: integer }] }
                machineId: { type: integer }
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
        "400": { description: invalid or unsupported mode }
        "501": { description: machine type does not support the settings proxy }
        "502": { description: Machine unreachable, timed out, or rejected the command }
  /api/machine/tare:
    post:
      tags: [MachineControl]
      summary: Request a scale tare (c_tare_pend over WebSocket)
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { machineId: { type: integer } } }
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
        "501": { description: machine type does not support the settings proxy }
        "502": { description: Machine unreachable, timed out, or rejected the command }
  /api/machine/service-test:
    post:
      tags: [MachineControl]
      summary: Briefly trigger a peripheral for the Maintenance page's component tests (c_service_test over WebSocket) — only works while the machine is idle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [peripheral]
              properties:
                peripheral: { description: "ServiceTestPeripheralDto enum name or numeric value", oneOf: [{ type: string, enum: [PUMP, VALVE, VALVE_B, LED] }, { type: integer }] }
                machineId: { type: integer }
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
        "400": { description: invalid peripheral }
        "501": { description: machine type does not support the settings proxy }
        "502": { description: Machine unreachable, timed out, or rejected (e.g. machine not idle) }
  /api/machine/profile/save:
    post:
      tags: [MachineControl]
      summary: Persist the currently active profile + its ID to flash (c_save_act_prof over WebSocket) — distinct from the saved-profile-slot CRUD under /api/machine/profile*
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { machineId: { type: integer } } }
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
        "501": { description: machine type does not support the settings proxy }
        "502": { description: Machine unreachable, timed out, or rejected the command }
  /api/machine/firmware/progress:
    get:
      tags: [MachineControl]
      summary: Poll an in-flight firmware OTA update's progress
      parameters:
        - { name: machineId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: "{ progress, status: IDLE|IN_PROGRESS|SUCCESS|ERROR, type: C_FW|F_FW|F_FS }"
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FirmwareProgress' }
        "501": { description: machine type does not support the settings proxy }
        "502": { description: Machine unreachable or returned an error }
  /api/machine/firmware/update:
    post:
      tags: [MachineControl]
      summary: Trigger a firmware OTA update from GitHub releases (ESP32 + STM32) — only starts the process, poll firmware/progress for status
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { machineId: { type: integer } } }
      responses:
        "200": { description: "{ message, success }" }
        "501": { description: machine type does not support the settings proxy }
        "502": { description: Machine unreachable or returned an error }
...